Начался рефакторинг

+++ lib/db_managers/directory_db_manager/include/DirectoryDbManagerMock.hpp
@@ -10,14 +10,14 @@ class DirectoryDbManagerMock : public IDirectoryDbManager {
 public:
     DirectoryDbManagerMock(std::shared_ptr<DbMock> db) ;
     
-    const Directory& get(size_t directory_id);
-    size_t add(const Directory& directory);
+    const Directory& get(int directory_id);
+    int add(const Directory& directory);
     void update(const Directory& directory);
-    void remove(size_t directory_id);
+    void remove(int directory_id);
 
 private:
     std::shared_ptr<DbMock> db_;
-    size_t aid_;
+    int aid_;
 
 };
