+++ lib/db_managers/node_db_manager/include/NodeDbManagerMock.hpp
@@ -1,26 +1,25 @@
 #pragma once
 
-#include <vector>
 #include <memory>
+#include <vector>
 
 #include "DbMock.hpp"
-#include "Node.hpp"
 #include "INodeDbManager.hpp"
 
-class NodeDbManagerMock : public INodeDbManager {
+class NodeDbManagerMock: public INodeDbManager{
 public:
-    NodeDbManagerMock(std::shared_ptr<DbMock> db);
+  NodeDbManagerMock(std::shared_ptr<DbMock> db);
 
-    const Node& get(size_t node_id);
-    size_t add(const Node& node);
-    void update(const Node& node);
-    void remove(size_t node_id);
+  const Node &get(int node_id);
+  int add(const Node &node);
+  void update(const Node &node);
+  void remove(int node_id);
 
-    std::vector<Node> getChildren(size_t node_id);
+//   void tag(int node_id, const Tag &);
+//   void move(int node_id, int destination_id);
+  std::vector<Node> getChildren(int node_id);
 
 private:
-    std::shared_ptr<DbMock> db_;
-    size_t aid_;
-
+  std::shared_ptr<DbMock> db_;
+  int aid_;
 };
-
