+++ lib/entities/calendar/include/Calendar.hpp
@@ -11,20 +11,18 @@ using CalendarSptr = std::shared_ptr<Calendar>;
 class Calendar {
  public:
   Calendar() = default;
-  // Для сохранения работоспособности некоторых тестов
-  // добавленные поля определены в конструкторе по умолчанию
   Calendar(
-      size_t id,
-      size_t node_id,
-      size_t owner_id,
+      int id,
+      int node_id,
+      int owner_id,
       std::string summary,
       std::string description,
       std::string visibility = std::string(),
       std::string color = std::string());
 
-  size_t id;
-  size_t node_id;
-  size_t owner_id;
+  int id;
+  int node_id;
+  int owner_id;
 
   std::string summary;
   std::string description;
