+++ lib/entities/calendar/src/Calendar.cpp
@@ -5,9 +5,9 @@
 #include <string>
 
 Calendar::Calendar(
-    size_t id,
-    size_t node_id,
-    size_t owner_id,
+    int id,
+    int node_id,
+    int owner_id,
     std::string summary,
     std::string description,
     std::string visibility,
