+++ lib/entities/directory/include/Directory.hpp
@@ -6,11 +6,11 @@
 class Directory {
 public:
     Directory();
-    Directory(size_t id, size_t node_id, size_t owner_id, std::string name, std::string description);
+    Directory(int id, int node_id, int owner_id, std::string name, std::string description);
 
-    size_t id;
-    size_t node_id;
-    size_t owner_id;
+    int id;
+    int node_id;
+    int owner_id;
 
     std::string name;
     std::string description;
