+++ lib/entities/event/src/Event.cpp
@@ -2,15 +2,15 @@
 #include <string>
 
 Event::Event(
-    size_t id,
-    size_t calendar_id,
+    int id,
+    int calendar_id,
     std::string summary,
     std::string description,
     Wt::WDateTime start,
     Wt::WDateTime end,
     bool is_recurrent,
     std::string frequency,
-    size_t interval,
+    int interval,
     Wt::WDate until)
   : id(id),
     calendar_id(calendar_id),
