+++ lib/entities/tag/include/Tag.hpp
@@ -6,9 +6,9 @@
 class Tag {
 public:
     Tag();
-    Tag(size_t id, std::string name);
+    Tag(int id, std::string name);
     
-    size_t id;
+    int id;
     std::string name;
 
 };
