+++ lib/entities/user/include/User.hpp
@@ -7,15 +7,15 @@ class User {
  public:
   User() = default;
   User(
-    size_t id,
-    size_t root_id,
+    int id,
+    int root_id,
     std::string login,
     std::string email,
     std::string nickname,
     std::string description);
 
-  size_t id;
-  size_t root_id;
+  int id;
+  int root_id;
 
   std::string login;
   std::string email;
