+++ lib/managers/calendar_manager/include/CalendarManager.hpp
@@ -13,16 +13,16 @@ class CalendarManager : public ICalendarManager {
 public:
     CalendarManager(std::shared_ptr<IDbManagers> db);
 
-    CalendarSptr get(size_t) override;
-    size_t add(CalendarSptr) override;
+    CalendarSptr get(int) override;
+    int add(CalendarSptr) override;
     void update(CalendarSptr) override;
-    void remove(size_t) override;
+    void remove(int) override;
 
-    std::vector<EventSptr> getEvents(size_t calendar_id) override;
+    std::vector<EventSptr> getEvents(int calendar_id) override;
     std::vector<EventSptr> getEventsByInterval(
-	size_t calendar_id,
-	Wt::WDateTime begin,
-	Wt::WDateTime end) override;
+        int calendar_id,
+	      Wt::WDateTime begin,
+	      Wt::WDateTime end) override;
 
 private:
     std::shared_ptr<IDbManagers> db_;
