+++ lib/managers/calendar_manager/include/ICalendarManager.hpp
@@ -9,17 +9,16 @@
 
 class ICalendarManager {
 public:
-	virtual CalendarSptr get(size_t calendar_id) = 0;
-	virtual size_t add(CalendarSptr calendar) = 0;  // Maybe returns error code 
-	virtual void remove(size_t calendar_id) = 0;
+	virtual CalendarSptr get(int calendar_id) = 0;
+	virtual int add(CalendarSptr calendar) = 0;  // Maybe returns error code 
+	virtual void remove(int calendar_id) = 0;
 	virtual void update(CalendarSptr calendar) = 0;
 
-	virtual std::vector<EventSptr> getEvents(size_t calendar_id) = 0;
+	virtual std::vector<EventSptr> getEvents(int calendar_id) = 0;
 	
 	virtual std::vector<EventSptr> getEventsByInterval(
-		size_t calendar_id,
-		Wt::WDateTime begin,
-		Wt::WDateTime end) = 0;
-
+      int calendar_id,
+		  Wt::WDateTime begin,
+		  Wt::WDateTime end) = 0;
 };
